+++ src/Hangfire.Mongo.Sample.ASPNetCore/Controllers/HomeController.cs
@@ -1,13 +1,23 @@
 ﻿using System;
 using System.Diagnostics;
 using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
+using Hangfire.Storage.Monitoring;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Hangfire.Mongo.Sample.ASPNetCore.Controllers
 {
     public class HomeController : Controller
     {
+        public class JobHandler
+        {
+            public void Execute(int index)
+            {
+                Thread.Sleep(200);
+                Debug.WriteLine($@"Hangfire fire-and-forget task started ({index}) - [{Thread.CurrentThread.ManagedThreadId}]");
+            }
+        }
         public ActionResult Index()
         {
             return View();
@@ -17,7 +27,7 @@ namespace Hangfire.Mongo.Sample.ASPNetCore.Controllers
         {
             Parallel.ForEach(Enumerable.Range(0, id), index =>
             {
-                BackgroundJob.Enqueue(() => PrintToDebug($@"Hangfire fire-and-forget task started ({index}) - {Guid.NewGuid()}"));
+                BackgroundJob.Enqueue<JobHandler>(j => j.Execute(index));
             });
 
             return RedirectToAction("Index");