+++ src/Hangfire.Mongo.Sample.ASPNetCore/Startup.cs
@@ -1,10 +1,9 @@
-﻿using Hangfire.Logging;
-using Hangfire.Logging.LogProviders;
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using MongoDB.Driver;
 using LogLevel = Hangfire.Logging.LogLevel;
 
 namespace Hangfire.Mongo.Sample.ASPNetCore
@@ -42,7 +41,7 @@ namespace Hangfire.Mongo.Sample.ASPNetCore
                 };
                 //config.UseLogProvider(new FileLogProvider());
                 config.UseColouredConsoleLogProvider(LogLevel.Trace);
-                config.UseMongoStorage(connectionString, "hangfire-mongo-sample-aspnetcore", migrationOptions);
+                config.UseMongoStorage(new MongoClient(connectionString), "hangfire-mongo-sample-aspnetcore", migrationOptions);
             });
             services.AddMvc();
 
@@ -51,7 +50,11 @@ namespace Hangfire.Mongo.Sample.ASPNetCore
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
         {
-            app.UseHangfireServer();
+            var options = new BackgroundJobServerOptions();
+            options.Queues = new[] {"default", "notDefault"};
+            options.WorkerCount = 2;
+            
+            app.UseHangfireServer(options);
             app.UseHangfireDashboard();
             if (env.IsDevelopment())
             {