+++ src/Hangfire.Mongo.Tests/MongoJobQueueFacts.cs
@@ -21,8 +21,9 @@ namespace Hangfire.Mongo.Tests
         {
             _jobQueueSemaphoreMock = new Mock<IJobQueueSemaphore>(MockBehavior.Strict);
             var queue = "default";
+            var timedOut = false;
             _jobQueueSemaphoreMock.Setup(s =>
-                    s.WaitAny(DefaultQueues, It.IsAny<CancellationToken>(), It.IsAny<TimeSpan>(), out queue))
+                    s.WaitAny(DefaultQueues, It.IsAny<CancellationToken>(), It.IsAny<TimeSpan>(), out queue, out timedOut))
                 .Returns(true);
         }
         [Fact]
@@ -116,7 +117,7 @@ namespace Hangfire.Mongo.Tests
                 connection.JobGraph.InsertOne(jobQueue);
                 var token = CreateTimingOutCancellationToken();
                 var queue = CreateJobQueue(connection);
-                _jobQueueSemaphoreMock.Setup(m => m.WaitNonBlock("default"));
+                _jobQueueSemaphoreMock.Setup(m => m.WaitNonBlock("default")).Returns(true);
                 
                 // Act
                 MongoFetchedJob payload = (MongoFetchedJob)queue.FetchNextJob(DefaultQueues, token);
@@ -151,7 +152,7 @@ namespace Hangfire.Mongo.Tests
                 connection.JobGraph.InsertOne(jobQueue);
 
                 var queue = CreateJobQueue(connection);
-                _jobQueueSemaphoreMock.Setup(m => m.WaitNonBlock("default"));
+                _jobQueueSemaphoreMock.Setup(m => m.WaitNonBlock("default")).Returns(true);
                 // Act
                 var payload = queue.FetchNextJob(DefaultQueues, CreateTimingOutCancellationToken());
 
@@ -190,7 +191,7 @@ namespace Hangfire.Mongo.Tests
                     FetchedAt = DateTime.UtcNow.AddDays(-1)
                 };
                 connection.JobGraph.InsertOne(jobQueue);
-                _jobQueueSemaphoreMock.Setup(m => m.WaitNonBlock("default"));
+                _jobQueueSemaphoreMock.Setup(m => m.WaitNonBlock("default")).Returns(true);
                 var queue = CreateJobQueue(connection);
 
                 // Act
@@ -238,7 +239,7 @@ namespace Hangfire.Mongo.Tests
 
                 var queue = CreateJobQueue(connection);
                 
-                _jobQueueSemaphoreMock.Setup(m => m.WaitNonBlock("default"));
+                _jobQueueSemaphoreMock.Setup(m => m.WaitNonBlock("default")).Returns(true);
 
                 // Act
                 var payload = queue.FetchNextJob(DefaultQueues, CreateTimingOutCancellationToken());
@@ -314,8 +315,8 @@ namespace Hangfire.Mongo.Tests
                 });
 
                 var queue = CreateJobQueue(connection);
-                _jobQueueSemaphoreMock.Setup(m => m.WaitNonBlock("critical"));
-                _jobQueueSemaphoreMock.Setup(m => m.WaitNonBlock("default"));
+                _jobQueueSemaphoreMock.Setup(m => m.WaitNonBlock("critical")).Returns(true);
+                _jobQueueSemaphoreMock.Setup(m => m.WaitNonBlock("default")).Returns(true);
                 
                 var critical = (MongoFetchedJob)queue.FetchNextJob(
                     new[] { "critical", "default" },